Fixed table definitions

+++ ion_auth.sql
@@ -5,7 +5,7 @@
 #
 
 CREATE TABLE `groups` (
-  `id` tinyint(3) unsigned NOT NULL AUTO_INCREMENT,
+  `id` mediumint(8) unsigned NOT NULL AUTO_INCREMENT,
   `name` varchar(20) NOT NULL,
   `description` varchar(100) NOT NULL,
   PRIMARY KEY (`id`)
@@ -28,7 +28,7 @@ DROP TABLE IF EXISTS `meta`;
 #
 
 CREATE TABLE `meta` (
-  `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
+  `id` mediumint(8) unsigned NOT NULL AUTO_INCREMENT,
   `user_id` int(10) unsigned DEFAULT NULL,
   `first_name` varchar(50) DEFAULT NULL,
   `last_name` varchar(50) DEFAULT NULL,
